--- sun/src/components/apexchart/index.tsx
@@ -1,3 +1,0 @@
-import dynamic from 'next/dynamic';
-
-export const ChartWrapper = dynamic(() => import('./chart'), { ssr: false });
